Moved geometry folder into plotters
