+++ scripts/01_directional_clustering.py
@@ -8,8 +8,6 @@ import numpy as np
 from compas.datastructures import Mesh
 
 # geometry helpers
-from compas.geometry import scale_vector
-from compas.geometry import dot_vectors
 from compas.geometry import subtract_vectors
 from compas.geometry import length_vector_sqrd
 
@@ -84,7 +82,7 @@ smooth_iters = 10  # how many iterations to run the smoothing for
 damping = 0.5  # damping coefficient, a value from 0 to 1
 
 # kmeans clustering
-clustering_name = "cosine kmeans" # algorithm name
+clustering_name = "variational kmeans" # algorithm name
 n_clusters = 5  # number of clusters to produce
 tol = 1e-6  # loss function threshold for early stopping
 iters = 30 # number of epochs to run kmeans clustering for
