+++ src/directional_clustering/plotters/geometry/lines.py
