+++ src/directional_clustering/plotters/geometry/vectors.py
